requirements.txt
attr==0.3.2
Brotli==1.1.0
brotlicffi==1.2.0.0
chardet==5.2.0
ConfigParser==7.2.0
cryptography==46.0.3
Faker==38.2.0
filelock==3.20.0
Flask==3.1.2
h2==4.3.0
HTMLParser==0.0.2
ipywidgets==8.1.8
mutagen==1.47.0
Pillow==12.0.0
pycryptodomex==3.23.0
pyOpenSSL==25.3.0
pywidevine==1.9.0
redis==7.1.0
simplejson==3.20.2
Sphinx==8.2.3
thread==2.0.5
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
wcwidth==0.2.14
yt_dlp==2025.11.12
zstandard==0.25.0
